test: Remove mock crate
