+++ bench-tx/Cargo.toml
@@ -16,8 +16,7 @@ path = "src/main.rs"
 [dependencies]
 miden-lib = { path = "../miden-lib", version = "0.4" }
 miden-objects = { path = "../objects", version = "0.4" }
-miden-tx = { path = "../miden-tx", version = "0.4" }
-mock = { package = "miden-mock", path = "../mock"  }
+miden-tx = { path = "../miden-tx", version = "0.4", features = ["testing"] }
 rand = { workspace = true }
 serde = { package = "serde", version = "1.0" }
 serde_json = { package = "serde_json", version = "1.0", features = ["preserve_order"] }
