+++ bench-tx/src/utils.rs
@@ -6,16 +6,9 @@ use miden_objects::{
     accounts::{Account, AccountCode, AccountId, AccountStorage, SlotItem, StorageSlot},
     assembly::ModuleAst,
     assets::{Asset, AssetVault},
-    notes::{Note, NoteId},
-    transaction::{ChainMmr, InputNote, InputNotes, OutputNote, TransactionArgs},
-    BlockHeader, Felt, Word,
+    Felt, Word,
 };
-use miden_tx::{DataStore, DataStoreError, TransactionInputs, TransactionProgress};
-use mock::mock::{
-    account::MockAccountType,
-    notes::AssetPreservationStatus,
-    transaction::{mock_inputs, mock_inputs_with_existing},
-};
+use miden_tx::TransactionProgress;
 use serde::Serialize;
 use serde_json::{from_str, to_string_pretty, Value};
 
@@ -44,109 +37,6 @@ pub const DEFAULT_ACCOUNT_CODE: &str = "
     export.basic_wallet::send_asset
     export.basic_eoa::auth_tx_rpo_falcon512
 ";
-
-// MOCK DATA STORE
-// ================================================================================================
-
-#[derive(Clone)]
-pub struct MockDataStore {
-    pub account: Account,
-    pub block_header: BlockHeader,
-    pub block_chain: ChainMmr,
-    pub notes: Vec<InputNote>,
-    pub tx_args: TransactionArgs,
-}
-
-impl MockDataStore {
-    pub fn new(asset_preservation: AssetPreservationStatus) -> Self {
-        let (tx_inputs, tx_args) =
-            mock_inputs(MockAccountType::StandardExisting, asset_preservation);
-        let (account, _, block_header, block_chain, notes) = tx_inputs.into_parts();
-
-        Self {
-            account,
-            block_header,
-            block_chain,
-            notes: notes.into_vec(),
-            tx_args,
-        }
-    }
-
-    pub fn with_existing(account: Option<Account>, input_notes: Option<Vec<Note>>) -> Self {
-        let (
-            account,
-            block_header,
-            block_chain,
-            consumed_notes,
-            _auxiliary_data_inputs,
-            created_notes,
-        ) = mock_inputs_with_existing(
-            MockAccountType::StandardExisting,
-            AssetPreservationStatus::Preserved,
-            account,
-            input_notes,
-        );
-        let output_notes = created_notes.into_iter().filter_map(|note| match note {
-            OutputNote::Full(note) => Some(note),
-            OutputNote::Partial(_) => None,
-            OutputNote::Header(_) => None,
-        });
-        let mut tx_args = TransactionArgs::default();
-        tx_args.extend_expected_output_notes(output_notes);
-
-        Self {
-            account,
-            block_header,
-            block_chain,
-            notes: consumed_notes,
-            tx_args,
-        }
-    }
-
-    pub fn tx_args(&self) -> &TransactionArgs {
-        &self.tx_args
-    }
-}
-
-impl Default for MockDataStore {
-    fn default() -> Self {
-        Self::new(AssetPreservationStatus::Preserved)
-    }
-}
-
-impl DataStore for MockDataStore {
-    fn get_transaction_inputs(
-        &self,
-        account_id: AccountId,
-        block_num: u32,
-        notes: &[NoteId],
-    ) -> Result<TransactionInputs, DataStoreError> {
-        assert_eq!(account_id, self.account.id());
-        assert_eq!(block_num, self.block_header.block_num());
-        assert_eq!(notes.len(), self.notes.len());
-
-        let notes = self
-            .notes
-            .iter()
-            .filter(|note| notes.contains(&note.id()))
-            .cloned()
-            .collect::<Vec<_>>();
-
-        Ok(TransactionInputs::new(
-            self.account.clone(),
-            None,
-            self.block_header,
-            self.block_chain.clone(),
-            InputNotes::new(notes).unwrap(),
-        )
-        .unwrap())
-    }
-
-    fn get_account_code(&self, account_id: AccountId) -> Result<ModuleAst, DataStoreError> {
-        assert_eq!(account_id, self.account.id());
-        Ok(self.account.code().module().clone())
-    }
-}
 
 // TRANSACTION BENCHMARK
 // ================================================================================================
