+++ miden-lib/src/tests/test_account.rs
@@ -6,38 +6,40 @@ use miden_objects::{
             ACCOUNT_ID_REGULAR_ACCOUNT_IMMUTABLE_CODE_ON_CHAIN,
             ACCOUNT_ID_REGULAR_ACCOUNT_UPDATABLE_CODE_OFF_CHAIN,
         },
+        testing::{
+            prepare_word, storage_item_0, storage_item_1, storage_item_2, storage_map_2,
+            transaction::{mock_executed_tx, mock_inputs, notes::AssetPreservationStatus},
+            MockAccountType, STORAGE_LEAVES_2,
+        },
         AccountId, AccountType, StorageSlotType,
     },
     crypto::{hash::rpo::RpoDigest, merkle::LeafIndex},
 };
-use mock::{
-    mock::{
-        account::{
-            storage_item_0, storage_item_1, storage_item_2, storage_map_2, MockAccountType,
-            STORAGE_LEAVES_2,
-        },
-        host::MockHost,
-        notes::AssetPreservationStatus,
-        transaction::{mock_executed_tx, mock_inputs},
-    },
-    prepare_transaction,
-    procedures::{output_notes_data_procedure, prepare_word},
-    run_tx, run_within_host, run_within_tx_kernel,
+use miden_tx::host::testing::{
+    procedures::output_notes_data_procedure,
+    utils::{prepare_transaction, run_tx, run_within_host, run_within_tx_kernel},
+    MockHost,
 };
 
 use super::{
     super::transaction::ToTransactionKernelInputs, ContextId, Felt, MemAdviceProvider,
     ProcessState, StackInputs, Word, ONE, ZERO,
 };
-use crate::transaction::memory::{ACCT_CODE_ROOT_PTR, ACCT_NEW_CODE_ROOT_PTR};
+use crate::transaction::{
+    memory::{ACCT_CODE_ROOT_PTR, ACCT_NEW_CODE_ROOT_PTR},
+    TransactionKernel,
+};
 
 // ACCOUNT CODE TESTS
 // ================================================================================================
 
 #[test]
 pub fn test_set_code_is_not_immediate() {
-    let (tx_inputs, tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
 
     let code = "
         use.miden::kernels::tx::prologue
@@ -67,7 +69,8 @@ pub fn test_set_code_is_not_immediate() {
 
 #[test]
 pub fn test_set_code_succeeds() {
-    let executed_transaction = mock_executed_tx(AssetPreservationStatus::Preserved);
+    let executed_transaction =
+        mock_executed_tx(AssetPreservationStatus::Preserved, &TransactionKernel::assembler());
 
     let output_notes_data_procedure =
         output_notes_data_procedure(executed_transaction.output_notes());
@@ -244,8 +247,11 @@ fn test_is_faucet_procedure() {
 #[test]
 fn test_get_item() {
     for storage_item in [storage_item_0(), storage_item_1()] {
-        let (tx_inputs, tx_args) =
-            mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+        let (tx_inputs, tx_args) = mock_inputs(
+            MockAccountType::StandardExisting,
+            AssetPreservationStatus::Preserved,
+            &TransactionKernel::assembler(),
+        );
 
         let code = format!(
             "
@@ -278,8 +284,11 @@ fn test_get_item() {
 
 #[test]
 fn test_set_item() {
-    let (tx_inputs, tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
 
     // copy the initial account slots (SMT)
     let mut account_smt = tx_inputs.account().storage().slots().clone();
@@ -333,8 +342,11 @@ fn test_set_item() {
 #[test]
 fn test_get_storage_data_type() {
     for storage_item in [storage_item_0(), storage_item_1(), storage_item_2()] {
-        let (tx_inputs, tx_args) =
-            mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+        let (tx_inputs, tx_args) = mock_inputs(
+            MockAccountType::StandardExisting,
+            AssetPreservationStatus::Preserved,
+            &TransactionKernel::assembler(),
+        );
 
         let code = format!(
             "
@@ -379,8 +391,11 @@ fn test_get_storage_data_type() {
 
 #[test]
 fn test_get_map_item() {
-    let (tx_inputs, tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
 
     let storage_item = storage_item_2();
     for (key, value) in STORAGE_LEAVES_2 {
@@ -427,8 +442,11 @@ fn test_set_map_item() {
         [Felt::new(9_u64), Felt::new(10_u64), Felt::new(11_u64), Felt::new(12_u64)],
     );
 
-    let (tx_inputs, tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
 
     let storage_item = storage_item_2();
 
@@ -482,8 +500,11 @@ fn test_set_map_item() {
 
 #[test]
 fn test_get_vault_commitment() {
-    let (tx_inputs, tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
 
     let account = tx_inputs.account();
     let code = format!(
@@ -513,8 +534,11 @@ fn test_get_vault_commitment() {
 
 #[test]
 fn test_authenticate_procedure() {
-    let (tx_inputs, _tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, _tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
     let account = tx_inputs.account();
 
     let proc0_index = LeafIndex::new(0).unwrap();
@@ -527,8 +551,11 @@ fn test_authenticate_procedure() {
     ];
 
     for (root, valid) in test_cases.into_iter() {
-        let (tx_inputs, tx_args) =
-            mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+        let (tx_inputs, tx_args) = mock_inputs(
+            MockAccountType::StandardExisting,
+            AssetPreservationStatus::Preserved,
+            &TransactionKernel::assembler(),
+        );
 
         let code = format!(
             "\
