+++ miden-lib/src/tests/test_faucet.rs
@@ -1,24 +1,27 @@
 use miden_objects::{
-    accounts::account_id::testing::{
-        ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1,
-        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
+    accounts::{
+        account_id::testing::{
+            ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1,
+            ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
+        },
+        testing::{
+            prepare_word,
+            transaction::{mock_inputs, notes::AssetPreservationStatus},
+            MockAccountType,
+        },
     },
     assets::FungibleAsset,
-};
-use mock::{
-    constants::{
-        non_fungible_asset, non_fungible_asset_2, CONSUMED_ASSET_1_AMOUNT, FUNGIBLE_ASSET_AMOUNT,
-        FUNGIBLE_FAUCET_INITIAL_BALANCE,
+    testing::{
+        constants::{
+            CONSUMED_ASSET_1_AMOUNT, FUNGIBLE_ASSET_AMOUNT, FUNGIBLE_FAUCET_INITIAL_BALANCE,
+        },
+        non_fungible_asset, non_fungible_asset_2,
     },
-    mock::{account::MockAccountType, notes::AssetPreservationStatus, transaction::mock_inputs},
-    prepare_transaction,
-    procedures::prepare_word,
-    run_tx,
 };
+use miden_tx::host::testing::utils::{prepare_transaction, run_tx};
 
 use super::ONE;
-use crate::transaction::memory::FAUCET_STORAGE_DATA_SLOT;
-
+use crate::transaction::{memory::FAUCET_STORAGE_DATA_SLOT, TransactionKernel};
 // FUNGIBLE FAUCET MINT TESTS
 // ================================================================================================
 
@@ -31,6 +34,7 @@ fn test_mint_fungible_asset_succeeds() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
     );
 
     let code = format!(
@@ -73,8 +77,11 @@ fn test_mint_fungible_asset_succeeds() {
 
 #[test]
 fn test_mint_fungible_asset_fails_not_faucet_account() {
-    let (tx_inputs, tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
 
     let code = format!(
         "
@@ -104,6 +111,7 @@ fn test_mint_fungible_asset_inconsistent_faucet_id() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
     );
 
     let code = format!(
@@ -135,6 +143,7 @@ fn test_mint_fungible_asset_fails_saturate_max_amount() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
     );
 
     let code = format!(
@@ -172,6 +181,7 @@ fn test_mint_non_fungible_asset_succeeds() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
     );
     let non_fungible_asset = non_fungible_asset(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN);
 
@@ -219,8 +229,11 @@ fn test_mint_non_fungible_asset_succeeds() {
 
 #[test]
 fn test_mint_non_fungible_asset_fails_not_faucet_account() {
-    let (tx_inputs, tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
     let non_fungible_asset = non_fungible_asset(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN);
 
     let code = format!(
@@ -246,8 +259,11 @@ fn test_mint_non_fungible_asset_fails_not_faucet_account() {
 
 #[test]
 fn test_mint_non_fungible_asset_fails_inconsistent_faucet_id() {
-    let (tx_inputs, tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
     let non_fungible_asset = non_fungible_asset(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1);
 
     let code = format!(
@@ -280,6 +296,7 @@ fn test_mint_non_fungible_asset_fails_asset_already_exists() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
     );
     let non_fungible_asset = non_fungible_asset_2(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN);
 
@@ -316,6 +333,7 @@ fn test_burn_fungible_asset_succeeds() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
     );
 
     let code = format!(
@@ -359,8 +377,11 @@ fn test_burn_fungible_asset_succeeds() {
 
 #[test]
 fn test_burn_fungible_asset_fails_not_faucet_account() {
-    let (tx_inputs, tx_args) =
-        mock_inputs(MockAccountType::StandardExisting, AssetPreservationStatus::Preserved);
+    let (tx_inputs, tx_args) = mock_inputs(
+        MockAccountType::StandardExisting,
+        AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
+    );
 
     let code = format!(
         "
@@ -391,6 +412,7 @@ fn test_burn_fungible_asset_inconsistent_faucet_id() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
     );
 
     let code = format!(
@@ -421,6 +443,7 @@ fn test_burn_fungible_asset_insufficient_input_amount() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::Preserved,
+        &TransactionKernel::assembler(),
     );
 
     let code = format!(
@@ -458,6 +481,7 @@ fn test_burn_non_fungible_asset_succeeds() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::TooManyNonFungibleInput,
+        &TransactionKernel::assembler(),
     );
     let non_fungible_asset_burnt = non_fungible_asset_2(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN);
 
@@ -512,6 +536,7 @@ fn test_burn_non_fungible_asset_fails_does_not_exist() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::TooManyNonFungibleInput,
+        &TransactionKernel::assembler(),
     );
     let non_fungible_asset_burnt = non_fungible_asset(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN);
 
@@ -546,6 +571,7 @@ fn test_burn_non_fungible_asset_fails_not_faucet_account() {
     let (tx_inputs, tx_args) = mock_inputs(
         MockAccountType::StandardExisting,
         AssetPreservationStatus::TooManyNonFungibleInput,
+        &TransactionKernel::assembler(),
     );
     let non_fungible_asset_burnt = non_fungible_asset_2(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN);
 
@@ -584,6 +610,7 @@ fn test_burn_non_fungible_asset_fails_inconsistent_faucet_id() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::TooManyNonFungibleInput,
+        &TransactionKernel::assembler(),
     );
     let non_fungible_asset_burnt = non_fungible_asset_2(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1);
 
@@ -625,6 +652,7 @@ fn test_get_total_issuance_succeeds() {
             empty_reserved_slot: false,
         },
         AssetPreservationStatus::TooManyNonFungibleInput,
+        &TransactionKernel::assembler(),
     );
 
     let code = format!(
