+++ miden-tx/Cargo.toml
@@ -18,8 +18,9 @@ path = "tests/integration/main.rs"
 
 [features]
 concurrent = ["miden-lib/concurrent", "miden-objects/concurrent", "miden-prover/concurrent", "std"]
-default = ["std"]
+default = ["std","testing"]
 std = ["miden-lib/std", "miden-objects/std", "miden-prover/std", "miden-verifier/std", "vm-processor/std"]
+testing = ["miden-objects/testing", "miden-lib/testing"]
 
 [dependencies]
 miden-lib = { path = "../miden-lib", version = "0.4", default-features = false }
@@ -30,5 +31,5 @@ rand = { workspace = true }
 vm-processor = { workspace = true }
 
 [dev-dependencies]
-mock = { package = "miden-mock", path = "../mock", default-features = false }
 rand_chacha = { version = "0.3", default-features = false }
+miden-tx = {path=".", features= ["testing"]}
