+++ miden-tx/src/executor/mod.rs
@@ -19,6 +19,9 @@ use crate::host::TransactionAuthenticator;
 mod data_store;
 pub use data_store::DataStore;
 
+#[cfg(feature = "testing")]
+pub mod testing;
+
 // TRANSACTION EXECUTOR
 // ================================================================================================
