+++ miden-tx/src/executor/testing/mod.rs
@@ -1,0 +1,126 @@
+// MOCK DATA STORE
+// ================================================================================================
+
+use alloc::vec::Vec;
+
+use miden_lib::transaction::TransactionKernel;
+use miden_objects::{
+    accounts::{
+        testing::{
+            transaction::{mock_inputs, mock_inputs_with_existing, notes::AssetPreservationStatus},
+            MockAccountType,
+        },
+        Account, AccountId,
+    },
+    assembly::ModuleAst,
+    notes::{Note, NoteId},
+    transaction::{
+        ChainMmr, InputNote, InputNotes, OutputNote, TransactionArgs, TransactionInputs,
+    },
+    BlockHeader,
+};
+
+use crate::{DataStore, DataStoreError};
+
+#[derive(Clone)]
+pub struct MockDataStore {
+    pub account: Account,
+    pub block_header: BlockHeader,
+    pub block_chain: ChainMmr,
+    pub notes: Vec<InputNote>,
+    pub tx_args: TransactionArgs,
+}
+
+impl MockDataStore {
+    pub fn new(asset_preservation_status: AssetPreservationStatus) -> Self {
+        let (tx_inputs, tx_args) = mock_inputs(
+            MockAccountType::StandardExisting,
+            asset_preservation_status,
+            &TransactionKernel::assembler(),
+        );
+        let (account, _, block_header, block_chain, notes) = tx_inputs.into_parts();
+        Self {
+            account,
+            block_header,
+            block_chain,
+            notes: notes.into_vec(),
+            tx_args,
+        }
+    }
+
+    pub fn with_existing(account: Option<Account>, input_notes: Option<Vec<Note>>) -> Self {
+        let (
+            account,
+            block_header,
+            block_chain,
+            consumed_notes,
+            _auxiliary_data_inputs,
+            created_notes,
+        ) = mock_inputs_with_existing(
+            MockAccountType::StandardExisting,
+            AssetPreservationStatus::Preserved,
+            account,
+            input_notes,
+            &TransactionKernel::assembler(),
+        );
+        let output_notes = created_notes.into_iter().filter_map(|note| match note {
+            OutputNote::Full(note) => Some(note),
+            OutputNote::Partial(_) => None,
+            OutputNote::Header(_) => None,
+        });
+        let mut tx_args = TransactionArgs::default();
+        tx_args.extend_expected_output_notes(output_notes);
+
+        Self {
+            account,
+            block_header,
+            block_chain,
+            notes: consumed_notes,
+            tx_args,
+        }
+    }
+
+    pub fn tx_args(&self) -> &TransactionArgs {
+        &self.tx_args
+    }
+}
+
+impl Default for MockDataStore {
+    fn default() -> Self {
+        Self::new(AssetPreservationStatus::Preserved)
+    }
+}
+
+impl DataStore for MockDataStore {
+    fn get_transaction_inputs(
+        &self,
+        account_id: AccountId,
+        block_num: u32,
+        notes: &[NoteId],
+    ) -> Result<TransactionInputs, DataStoreError> {
+        assert_eq!(account_id, self.account.id());
+        assert_eq!(block_num, self.block_header.block_num());
+        assert_eq!(notes.len(), self.notes.len());
+
+        let notes = self
+            .notes
+            .iter()
+            .filter(|note| notes.contains(&note.id()))
+            .cloned()
+            .collect::<Vec<_>>();
+
+        Ok(TransactionInputs::new(
+            self.account.clone(),
+            None,
+            self.block_header,
+            self.block_chain.clone(),
+            InputNotes::new(notes).unwrap(),
+        )
+        .unwrap())
+    }
+
+    fn get_account_code(&self, account_id: AccountId) -> Result<ModuleAst, DataStoreError> {
+        assert_eq!(account_id, self.account.id());
+        Ok(self.account.code().module().clone())
+    }
+}
