+++ miden-tx/src/host/tx_authenticator.rs
@@ -166,8 +166,8 @@ impl TransactionAuthenticator for () {
 
 #[cfg(test)]
 mod test {
+    use miden_lib::utils::{Deserializable, Serializable};
     use miden_objects::{accounts::AuthSecretKey, crypto::dsa::rpo_falcon512::SecretKey};
-    use mock::utils::{Deserializable, Serializable};
 
     #[test]
     fn serialize_auth_key() {
