--- mock/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "miden-mock"
-version = "0.4.0"
-description = "Library for mocking and testing core Miden rollup components"
-readme = "README.md"
-keywords = ["miden", "mock"]
-license.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-edition.workspace = true
-
-[[bin]]
-name = "mock"
-path = "src/main.rs"
-required-features = ["executable"]
-
-[features]
-default = ["std"]
-executable = ["std", "serde", "dep:clap"]
-serde = ["dep:serde", "miden-objects/serde"]
-std = ["miden-lib/std", "miden-objects/std"]
-
-[dependencies]
-clap = { version = "4.4", features = ["derive"], optional = true }
-env_logger = { version = "0.11" }
-hex = { version = "0.4" }
-miden-lib = { path = "../miden-lib", version = "0.4" }
-miden-objects = { path = "../objects", version = "0.4", features = [
-    "log",
-    "serde",
-    "testing",
-] }
-miden-prover = { workspace = true }
-postcard = { version = "1.0", features = ["alloc"] }
-rand = { version = "0.8" }
-rand-utils = { package = "winter-rand-utils", version = "0.8" }
-rand_pcg = { version = "0.3", features = ["serde1"] }
-serde = { version = "1.0", optional = true, default-features = false, features = [
-    "derive",
-] }
-vm-processor = { workspace = true, features = ["internals"] }
