--- mock/src/mock/mmr_peaks.rs
@@ -1,7 +1,0 @@
-use alloc::vec::Vec;
-
-use miden_objects::crypto::merkle::MmrPeaks;
-
-pub fn empty_mmr_peaks() -> MmrPeaks {
-    MmrPeaks::new(0, Vec::new()).unwrap()
-}
