--- mock/src/mock/mod.rs
@@ -1,9 +1,0 @@
-pub mod account;
-pub mod block;
-pub mod chain;
-pub mod host;
-pub mod mmr_peaks;
-pub mod notes;
-pub mod transaction;
-
-use super::builders;
