+++ objects/Cargo.toml
@@ -24,13 +24,14 @@ concurrent = ["std"]
 default = ["std"]
 serde = ["dep:serde", "miden-crypto/serde"]
 std = ["assembly/std", "miden-crypto/std", "miden-verifier/std", "vm-core/std", "vm-processor/std"]
-testing = ["dep:winter-rand-utils"]
+testing = ["dep:winter-rand-utils", "dep:rand"]
 
 [dependencies]
 assembly = { workspace = true }
 log = { version = "0.4", optional = true }
 miden-crypto = { workspace = true }
 miden-verifier = { workspace = true }
+rand = { workspace = true, optional = true }
 serde = { version = "1.0", optional = true, default-features = false, features = ["derive"] }
 vm-core = { workspace = true }
 vm-processor = { workspace = true }
