+++ objects/src/accounts/storage/mod.rs
@@ -12,6 +12,9 @@ pub use slot::StorageSlotType;
 mod map;
 pub use map::StorageMap;
 
+#[cfg(feature = "testing")]
+pub mod testing;
+
 // CONSTANTS
 // ================================================================================================
