+++ objects/src/assets/mod.rs
@@ -18,6 +18,9 @@ pub use token_symbol::TokenSymbol;
 mod vault;
 pub use vault::AssetVault;
 
+#[cfg(feature = "testing")]
+pub mod testing;
+
 // ASSET
 // ================================================================================================
